secboot: add workaround for snapcore/core-initrd issue #13
This commit works around snapcore/core-initrd#13, i.e. we cannot run
```
systemd-run --wait --pipe
```
in the initrd because there is no dbus there and those features need
dbus.

So this PR implements a workaround that uses polling to get the result
and tmp files. Not nice but workable.

* secboot/secboot_tpm.go: re-add useful comment

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* secboot/secboot_tpm.go: increase frequency we check for hook completion

The resolution of how often we check directly affects boot speed, since we need
to wait at least one cycle if the file is not created instantly, so rather than
always hold up boot for at least one second, make the time period between checks
smaller but not too small such that we are hogging CPU cycles checking for a
file. 20 Hz seems a reasonable compromise for this.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* secboot/secboot_tpm.go: manually delete and recreate the std{in,out,err} streams

We want to be robust and secure, so to guard against possible bugs where the
files weren't properly deleted and/or have the wrong permissions, we always
should delete and re-create the stream files with the right permissions before
executing the hook itself.

Also add a test with a hook implementation which verifies the files have the 
right permissions and content that we expect.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* secboot/secboot_tpm_test.go: fix typo

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

Co-authored-by: Ian Johnson <ian.johnson@canonical.com>