o/devicestate/firstboot20_test.go: test for different model grades
We will be getting more code which is model grade specific to devicestate and
snapstate, so it's useful to test with different grades here. For the existing
tests just running them all with different model grades and expecting the same
behavior is good enough, eventually we will have separate grade specific tests
however.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>