snap: skip help output tests for go-flags v1.4.0
Skip --help output tests for older versions of
go-flags. Notably v1.4.0 from debian-sid will fail because
the formating is slightly different. Note that the check here
is not precise i.e. this is not the commit that added the change
that changed the help output but this change is easy to test for
with reflect and in practice this is fine.