cmd/snap-repair: filter assertions by bases and modes
Use the bases and modes settings in repair assertions to filter repair
assertions when snap-repair is running.

The logic for modes is such that when modes is not set, on UC20 devices,
the assertion only applies to run mode devices. If modes include "run",
non-UC20 devices will also be included. To filter just on UC20 or UC18
devices, the bases setting should be set in assertion.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>