gadget: enable multi-volume uc20 gadgets in  LaidOutVolumeFromGadget; rename too
Rename LaidOutVolumeFromGadget to LaidOutUbuntuVolumeFromGadget, since now the
function returns the specific volume with all the ubuntu roles on it.
Additionally, allow multiple volumes and choose the right one by searching for
the SystemSeed role/partition.

Also move the old implementation of LaidOutVolumeFromGadget to a test-only
helper in the gadget/install package, since there are many tests in the install
package which use this helper with invalid gadget.yaml's that don't contain any
ubuntu roles and porting all of those tests to also include ubuntu roles is a
bit of a fool's errand.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>