gadget: pass sector size in to mkfs family of functions, use to select block sz
If we are creating partitions that are small such as ubuntu-save, then we have
code to specify the block size, but on some devices where the sector size is
not the default of 512 (and is instead 4K for example), we need to adjust the
block size when creating the filesystem to something that is at least the same
as the sector size. This is confirmed looking at the source of e2fsprogs at
https://github.com/tytso/e2fsprogs/blob/master/misc/mke2fs.c#L2151-L2156.

As such, pass the sector size along to where we call mkfs, and adjust the
specified block size to be the same as the sector size if the sector size is
larger than 1K when creating ext4 and vfat partitions.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>