bootloader/assets: support injecting bootloader assets for tests
Add support for injecting bootloader assets and corresponding snippets in test
builds of snapd. The code is only picked up when building with
withbootassetstesting tag.

The snippets are injected only when a specific marker file is present in the
same directory as the snapd binary. This way it is possible to control when a
snapd snap update will trigger an update of boot config assets of the grub
bootloader.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>