gadget/gadget.go: deny nil models in LaidOutSystemVolume..., rm label option
The role check here is sufficient, provided that we provide a non-nil model to
ReadInfoAndValidate, since with a valid model that function will promote
structures that for example only have a filesystem-label: system-boot, but don't
have a role: system-boot in them to also have the role: system-boot set.

This function in practice is always meant to be used with a non-nil model, so to
protect against this function failing to find the system-boot role structure,
prevent calling it with a nil model. Also test for this too.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>