snap/info.go: ignore unknown daemons in SortSnapServices
When a snap has daemons that use before/after, and a hook tries to specify
snapctl restart of a single daemon, then the code here is subtly wrong in that
it thinks that there is an ordering cycle because we couldn't dequeue the other
services since they don't exist in the list of services provided to
SortSnapServices. As such, we should only sort the services that actually exist
in the list we were provided, we will handle errors around unknown services that
don't exist at all in the snap elsewhere in the stack before we get here.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>