cmd/snap-bootstrap/initramfs-mounts: write realistic modeenv for recover+install
The current modeenv that is written from the initramfs is incomplete to the
extent that it doesn't accurately reflect what base and model are booted on the
system so when things like snap-repair try to determine device information using
only the modeenv, they fail.

This adds as much information as can be trivially gained from the initramfs
about the device/system. The current kernel snaps, trusted boot assets and 
kernel command lines can probably also be determined at this point and included
in the modeenv as well, but it's unclear if that's the direction we want to go
with the modeenv we write here.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>