osutil: skip TestReadBuildGo inside sbuild
The TestReadBuildGo test started to fail recently inside an
hirsute build environment with:
```
----------------------------------------------------------------------
FAIL: buildid_test.go:131: buildIDSuite.TestReadBuildGo

buildid_test.go:142:
    c.Assert(string(output), Equals, "")
... obtained string = "" +
...     "# command-line-arguments\n" +
...     "loadinternal: cannot find runtime/cgo\n"
... expected string = ""
```

Failure reason is unknown but only reproducible inside the particular
a dpkg-buildpackage build on 21.04 (with dh-golang).

Not reproducible outside the dpkg-buildpackage when running tests
in a normal hirsute environment.

This is a first stop-gap to fix the FTBFS until the root cause is
found.