snap: rename gdbserver option to `snap run --gdbserver`
The `snap run --experimental-gdbserver` option is available
since a while and we got no negative feedback so far. Given
that it's also superior to the `snap run --gdb` option this
PR make it available with the *experimental* prefix.

The old `snap run --experimental-gdbserver` option is still
available for compatbility and can be removed after some
releases.