tests/lib/fakestore/store/store.go: handle If-None-Match and unsupported Accept
If the Accept header is something other than application/json or
application/x.ubuntu.assertion, then we should return a 400 to the client.

The If-None-Match header as used by snap-repair is essentially a caching
mechanism insofar as if the If-None-Match setting is equal to the revision of
the assertion we found, then we don't have to return anything and can just set
the status code to 304 and don't need to put anything in the Body of the
response.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>