o/snapstate/check_snap.go: fix off by 1 loop range error
Off by 1 strikes again! The previous code was written such that req and cur had
the full string/regex match as the first element, and so the loop needed to skip
that first element in the comparison, but here now req and cur have just the
elements of the version number in them, so we need to include the first element
(AKA index 0) in the comparison.

Luckily this was actually caught by spread tests, where we have the major
version of snapd as 1337.2.49, but a snap like docker which we test installing
uses an assume on snapd2.40, which the parsing code was ignoring the fact that
1337 > 2, and instead started with the comparison 2 < 40 and failed the check.

So also add some unit tests which test this case. Ideally also a spread test
eventually too that doesn't require docker in case that snap goes away.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>