o/configstate/configcore: support snap set system swap.size=...
Add a handler for the system swap option which manipulates /etc/default/swapfile
and swapfile.service. This option works both with systemd around at runtime, and
without during the initrd for example when only the rootfs is available.

In order to properly handle this when the key is not set without running systemd
unnecessarily, we need to use GetPristine to indicate if the value was changed
or not.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>