boot/flags.go: add RunModeRootfs()
This function is meant to be used by snap{,ctl} system-mode in reporting the
location of ubuntu-data as it is used by run mode. For recover mode in order to
find the location of ubuntu-data, we need to parse degraded.json to get the
mount-location and mount-state, but for the other modes the directory is
statically known.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>