tests/lib/prepare: fix repacking of the UC20 kernel snap for with ubuntu-core-initramfs 40
Version 40 of ubuntu-core-initramfs changed how the --kerneldir command line
parameter is treated and it is expected to point to the actual kernel version
modules directory.

The version of u-c-i also requires kernel firmware to be present, but
unfortunately does not declare such dependency and we need to install the
package ourselves.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>