tests/lib/prepare: workaround u-c-i to use the current skeleton for new initrd
We want to use the skeleton rootfs from the old initrd, rather than the one that
comes from the current release of ubuntu-core-initramfs package. Workaround some
assumptions in u-c-i and pass 'main' as a feature, which is the name of the
directory inside --skeletondir we use, to make it copy that directory and use it
as a template.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>