interfaces/apparmor: allow reading /proc/sys/kernel/random/entropy_avail
Allow reading how much entropy is available in the kernel. This is a
miscellaneous access that some applications, eg. those that generate random data
may need. Note that the AppArmor base abstraction already allows reading from
/dev/{u,}random, and our base template allows writing to it. Thus, even without
knowing how much entropy is available, it is possible to drain it.

See [1] for more details.

1. https://forum.snapcraft.io/t/allow-access-to-proc-sys-kernel-random-entropy-avail/23275

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>