snapstate: reduce reRefreshRetryTimeout to 1/2 second
We got some reports that snapd during download consumes a lot of
CPU. It turns out this was introduced with PR#6356 when we add
re-refresh on epoch bumps. This commit mitigates the effect.

Fixing this for good will require some work in the taskrunner.
But with this change I see on my machine that download only
takes ~14% compared to the 56% it took before.