tests/nested/manual: test that uc20 initramfs moves time with broken RTC
This ensures that when folks build a new image/model assertion with a brand key
that was created after the kernel initrd on a device without a RTC, the time
will be moved forward. We set both the nested VM's RTC and the kernel initrd
timestamp to the unix epoch such that the initramfs _needs_ to move the time
forward in order for assertion handling to be verified correctly now.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>