o/{device,hook}state: encode fde-setup-request key as base64 string
The spec states that the encryption key in the fde-setup-request JSON structure
that is passed to the hook via `snapctl` should be a string encoded as base64.

We missed the fact that this code was not doing that, since it was refactored
rather late in the cycle, and the existing test (and also the real world hook)
are both written in Go, and thus take advantage of Go's advanced JSON decoding
which handles both the case of

```json
{
   ...
   "key": [96, 34, 56, 54, ... ],
   ...
}
```

```json
{
   ...
   "key": "ABCDE==",
   ...
}
```

the same when decoding. Actually because of this fact, we can in fact change the
request we send to conform to the spec without breaking the existing client and
test.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>