tests/lib/fde-setup-hook: also verify that fde-reveal-key key data is base64
The key data should be decodable both as a []byte in Go directly (using the
standard library's automatic conversion of base64 encoded strings into []byte),
as well as manually by reading the base64 string and decoding out of band from
the Go standard library.

We added code to do this for the fde-setup-request, but we should also do this
check for the fde-reveal-key structure too just to be extra safe around this
area.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>