gadget: resolving one reference from kernel.yaml in gadget is enough
This changes the policy so that resolving one kernel assets in
the gadget is enough enough. E.g.
If the kernel declares it provides "ref1", "ref" and the gadget
has a "$kernel:ref1" OR "$kernel:ref2" in gadget.yaml that is
considered fine. The rational is that the kernel could provide
assets for different boards and the gadget would then pick
from the kernel what it needs.