devicestate: drop passing `Models` in the fde.SetupRequest
As outlined in the most recent version of the fde-helper
interface we will no longer pass the models to the hook. The
upper layers of secboot will care about binding the fde key
to the models.

None of the existing consumers of the fde-hooks are using
the models so it's safe to drop them.