wrappers/services.go: introduce EnsureSnapServices()
This function will modify snap service files only as needed, if none of the
snap service files need to be updated, then nothing is done, otherwise just the
service files which don't exist or need to be modified are actually written.

It also properly handles rollbacks, where the previous state before the
function runs is saved and restored to if anything does happen to go fatally
wrong.

Also add some tests for the new behavior of EnsureSnapServices around rollbacks
which will be needed for certain features we add to snapd.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>