wrappers/services.go: refactor options handling for Service generation
Introduce SnapServiceOptions, and make EnsureSnapServices accept this as the
value of the map argument.

Also add RequireMountedSnapdSnap to EnsureSnapServicesOptions because this is
in fact a global property that should apply to all snap services that are being
ensured.

These two changes taken together mean we can eliminate ambiguity as to which
options are used in EnsureSnapServices and make it much more user friendly with
no option duplication between the map argument and the global options argument.

Also add some more unit tests around the options handling to ensure that global
options are applied properly to different snaps when those snaps have different
per-snap options.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>