tests/lib/nested: do not compress images
Stop compressing the images as it seems to be a waste of time. Analyzing some
UC20 runs, the image built by ubuntu-image is 199MB (10GB apparent size, since
those are sparse files). Compressing that image with xz -0 takes ~2 minutes,
inflating it takes another ~30s. The compressed image is 195MB (vs 199MB
initially).

Instead of compression, use simple cp to keep track of the pristine versions of
the images and preserve the sparse file property.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>