snap/quota: support orphaning and adopting groups with ResolveCrossReferences
This is needed at least so that we can create separate groups and add them to
another existing group. We need to be careful that internal links are managed
appropriately and also double validated in the case that a group is too big is
attempted to be adopted by a group which cannot fit the adoptee inside it.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>