o/snapstate: remove snap from quota groups on discard
When removing or disabling a snap, we should remove references to it from the
quota the snap was in before it was removed so that the quota group it was in
remains in a valid state.

Note that this is only executed for snaps that are removed, if the snap is just
disabled, it remains in the state for that quota since it could be re-enabled
at some point in the future.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>