sandbox/cgroup: wait for pid to be moved to the desired cgroup
Through experimentation, it was determined that the requesting systemd to move a
process to a cgroup by exchanging messages over dbus, returns a success before
the change is observable by inspecting the process' /proc/<pid>/cgroup data.

Add a little delay and wait for up to 100ms for the process to be moved.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>