release-tools/changelog.py: handle case where we don't have a TZ
Somehow, on my system the %z format string was empty, resulting in the empty
string being used for the timezone, which does not match the expected format
for the opensuse changelog spec file. In this case, if we detect that there is
no timezone, use utc time and use the +0000 as the timezone.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>