snap/quota: check that current tasks count is non-zero before memory usage
This avoids a bug on CentOS 7 where the current memory usage for a slice which
is active but has no tasks in it is reported as 16 exbibytes, which overflows
integers and breaks our parsing logic in the systemd logic. For these slices,
systemd does report correctly that there are no processes in the slice, so we
can thus deduce that there is no memory usage for such a slice and avoid the
bug.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>