snap/quota: remove workaround for CurrentMemoryUsage
This workaround was for broken systemd's in an attempt to still support these
systems, but the bugs ended up overtaking us and the situation became
untenable.

So remove the workaround so we just check if the slice is active and if it is
return whatever systemd says, with the understanding that we should never
execute this code on broken systemd versions.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>