tests/snap-quota-groups: wait for svc to be in failed SubState
After we create the quota group, we will attempt to start the service, which
will then fail due to OOM, but the default behavior of snap services is to use
restart-condition: always, which means that systemd attempts to restart failed
services infinitely until of course the service hits the StartLimitBurst, at
which point then we should check that the service is inactive.

If we check the state of the service before systemd has finished trying to
restart it, then we risk seeing it active since it will be in state "starting"
or "running" or "auto-restart", none of which are inactive.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>