interfaces/apparmor: allow snap-update-ns to work with v2 freezer
When using unified cgroups chierarchy, the snap processes are placed in a
tracking cgroup. This is either arranged by systemd for services or by snap run
for user applications in which case they end up in a scope.

However, the tracking group can have a random suffix, thus when freezing all
processes from a given snap, snap-update-ns must walk the tree to find the
matching groups. And only then act on each of them.

In the process, the snap-confine and snap-update-ns chain has already been put
inside the trackign cgroup of a given snap. Thus our own group must be
determined such that it can be skipped when freezing all other groups.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>