cmd/snap-confine: make mount ns use check cgroup v2 compatible
Fix the mount namespace usage check to work with a cgroup v2 setup. Previously,
the snap processes would end up in a separate group under the freezer
controller, making the check simple and efficient.

With cgroups v2, there is are separate controllers. However, each snap process
is moved to a tracking group (systemd scope) created by snap run early in the
call chain. We can use that property to locate processes from a given snap. The
groups are managed externally by systemd, which takes care of the cleanup once
the last process exits. A slight detour is that snap-confine is no longer
responsible for placing the process inside the group, as the step has been done
early on by either the system or user systemd instance.

The patch refactors the code to be aware of the tracking group, and moves
through the group hierarchy to find one which is named after a snap. In the
process, we must be aware of our own group, such that it can be skipped.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>