many: use changes + tasks for quota group operations
The exported methods from the servicestate package all return tasksets now,
which is expected to be put into changes that are executed by the overlord
loop. This involves changes in many parts that use quotas, such as the tests,
where a new mock function to create a quota group in state without running any
servicectl commands as well as changes to the client side.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>