packaging/{debian-sid,ubuntu-14.04}/snapd.postrm: remove slices too
It was missed that we don't remove slices on debian sid and trusty before, so
remove those here too. This is slightly overkill on trusty, since the version
of systemd there is too old to properly support quota groups which generates
the slices, but remove them anyways if they have the X-Snappy=yes since we
wrote them there at some point somehow.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>