tests/lib/store.sh: fix make_snap_installable_with_id()
The make_snap_installable_with_id() was not actually copying the snap
binary into the fake store archive, because that's a step performed by
the new_snap_declaration() function, which was not invoked here.

By using that function we ensure that the snap binaries are made
available to the store. It also automatically acks the assertions.

For consistency, we also use new_snap_revision(), which apart from the
extra "ack" on the assertion is equivalent to the previous
implementation.