configcore: ignore system.pi-config.* setting on measured kernels
* configcore: ignore system.pi-config.* setting on measured kernels

Some pi devices uses measured boot, e.g. avnet. On these devices
the config.txt is part of the measurements and we cannot modify
config.txt without changing the measurements and breaking the
boot.

So this commit adds a check for kernels that known to use measured
boot and here the configuration is ignored.

* configcore: tweak naming/comments (thanks to Samuele)