snapstate: remove temporary snap file for local revisions early
* overlord: add manager tests for refresh to local revisions

This adds manager tests to test that refresh to the same revision
works and local revisions work.

* snapstate: remove temporary snap file for local revisions early

The SnapSetup.SnapPath is only needed to check/mount the snap. If
the snap is already local we will not mount it again. So we can
remove temporary snap files from e.g. side-loading early.

This fixes the bug that sideloaded snaps that already have local
revisions are not removed. They will get removed eventually via
the snapmgr.go:SnapManager.localInstallCleanup() code. But this
code only runs every 24h and if people install enough side-loaded
snaps the disk may fill-up before the cleanup happens.

* snapstate: move removal of local revisions closer to the "mount" task generation (thanks to Samuele)

* snapstate: fix typo