overlord: only start enabled services (#10438)
* overlord/servicestate: load list of disabled services

When starting services, read the list of disabled services from the
state changes structure.

* wrappers/services: make querying of active services more generic

Instead of taking the snap info structure, only takes a map of the
snap's services. This makes these functions easier to reused in other
circumstances.

* overlord/servicestate: compute list of disabled services

When processing the "snap start" command, compute the list of disabled
services. We don't do it if the `--enable` flag is also given, because
in that case we don't care about the actual state of the services.

* overlord/servicestate: always start explicitly mentioned services

Do not add the explicitly mentioned services to the disabledServices
parameter, as that would cause them not to get started.

* tests: add spread test for "snap start" with disabled services

* overlord: delay computation of disabled services

* tests: make tests robust against re-ordering

+ address review comments
   tests: stop only service 3