o/devicestate, sysconfig: refactor cloud-init config permission handling
Move more responsibility of handling cloud-init config permission handling to
sysconfig instead of devicestate since we now have access to the model in
ConfigureTargetSystem which we can use to get the model grade.

This also means we can refactor the functions which install cloud-init config
to be more straight forward and readable.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>