tests/nested/manual: use --collect with systemd-run always with --unit
When we specify --unit in a test, trying to run that test twice will fail
because the old unit will still exist (even if it is inactive/stopped), so the
next iteration of the test systemd-run cannot create a new unit with the
specified name. As such, we can specify --collect such that when the unit is
stopped, it will disappear such that we can re-create it on future iterations
of the test.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>