interface/modem-manager: add support for MBIM/QMI proxy clients
The mbim-proxy daemon is started by the ModemManager service.
The modem's app which wants to send MBIM messages will
connect to the mbim-proxy daemon via the unix socket.

The mbim-proxy daemon uses the the abstract socket path
"mbim-proxy". The modem-manager interface needs a
corresponding update. This provides that update.

References:
https://bugs.launchpad.net/austin/+bug/1936374