configcore: fix early config timezone handling
There was a regression from PR#10562 that added the virtual timezone
handling. When the timezone is set from the config defaults it will
be applied in "install" mode and when the system then installs to
the target disk it will set the timeone again, this time with the
fsOnlyContext set. When this happend the current timezone and the
new timezone are identical and the code would do nothing. This is
incorrect however, the check if the system timezone and the new
timezone matches only makes sense when setting the "runtime"
system, not the target filesystem.

This commit fixes this issue by moving the check into the right
place.