packaging: remove TEST_GITHUB_AUTOPKGTEST support
In the (very) old days before spread we ran our testsuite in the
autopkgtest infrastructure. This had many problems and when spread
came along we used that and never looked back.

The code to run the tests in autopkgtest is still in place [1] and
this commit removes it now as it is unlikely we ever us this
again (and if so we can always revert this commit).

[1] In debian/rules because autopkgtest really only can deal with
debian packages.