.github/workflows/test.yaml: fix boolean logic
We want the spread tests to also run on pushes to release/** branches, so the
operator here should just be || otherwise these jobs will not run on pushes.
Thanks to Alberto for pointing this out to me.

Furthermore, we don't want to run the snap-builds job on anything other than
pull requests, since we already have Launchpad snap build recipes setup for the
release and master branches and this would just be duplicating that effort and
use up artifact storage space.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>