tests: cherry-pick shellcheck fix `bd730fd4` from #10443 (2.51)
The new and the older shellcheck versions disagree about the
nature of the echo -e warning. New shellcheck warns with SC3037
and old with SC2039. The old shellcheck runs when in the
tests/unit/go spread test and the new in GH actions. Given that
this is the only instance where this is a problem this commit just
makes shellcheck ignore both the old and the new code for the issue.

This was merged as part of #10443 [1] but because it got squashed
we need the isolated fix for 2.51.

[1] https://github.com/snapcore/snapd/pull/10443/commits/bd730fd49cbb0309430c5fa3fecc35073d2d343d