wrappers: fix a bunch of duplicated service definitions in tests
We have a bunch of duplicated key in our snap.yaml that we use in
tests. This is not allowed with the yaml spec but `yaml.v2` is
leaniant about it. It becomes a problem with `yaml.v3` because
that errors for duplicated keys. Hence this first commit to fix some
obvious duplication. There will be more :)