tests/lib/prepare.sh: add debug kernel command line params via gadget on UC20
This aids greatly in debugging failure to boot in GCE when there are problems
like not being able to boot in GCE, but QEMU works fine, since we can view the
serial console on GCE via some settings but can't view i.e. tty1.

This requires adjusting the gadget-update-pc spread test, which was written for
the assumption that the pc snap would always be asserted, which is no longer
true, so we have to use --dangerous on UC20 for this test.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>