packaging: ship the `snapd.apparmor.service` unit in debian
The `snapd.apparmor.service` was not needed in debian because
it used a version of apparmor that would do the snapd apparmor
profile loading for us. However this support was dropped with
apparmor 3 and when it entered unstable this broke snapd.

This commit ensures it is shipped on debian too.