sysconfig/cloudinit: fix datasource restriction file installation on dangerous
With grade dangerous we don't want to install the datasource restriction file,
so only install it if we installed files and if we are filtering those files.

In the case where we are filtering files but there are no files to install, we
will already not create the file, but if we are filtering and we try to install
some files but they all get filtered out, then we will not install a datasource
restriction file because nothing untrusted from ubuntu-seed was installed, only
files from the gadget which are never filtered but are trusted.

Also remove a redundant test case now that it is reflected in the bigger test
case table.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>