cmd/libsnap-confine-private: cgrupv2 die when process is not in snap specific group
It is possible that the snap process has not been moved to the snap specific
tracking cgroup. In this case, setting up a device filtering on the group can
negatively affect whatever group the process is part of. Try to catch test
scenarios when thiss happens, so that we may reach a reasonable solution.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>