osutil/disks: add Disk.FindMatchingPartitionWith{Fs,Part}Label
These methods were already implemented for *disk, so expose them as part of the
interface so they can be directly tested.

Also, temporarily implement these methods for the MockDiskMapping using the
existing mappings that are provided with a TODO to switch over the
implementation to searching the static list when that static list is provided
by the test callers that need to mock this.

Finally notate in the doc-comments of
FindMatchingPartitionUUIDWith{Fs,Part}Label that these methods will be taken
out to pasture soon.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>