gadget/many: filesystemInfo -> ...InfoForPartition, unify tests' lsblk mocking
Rename and refactor the helper filesystemInfo to be specifically for
partitions, including to the point where we now call lsblk twice, once to
verify that the device passed to us is indeed a partition, and then the second
time to get the filesystem info for that partition.

Since this requires rather extensive changes to lsblk testing, unify all the
mocking of that command to use a common shared helper now located in the new
gadgettest package.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>