osutil/disks: add DiskFromPartitionDeviceNode
This will be used when we have /dev/sda1 for example and want to get a Disk of
/dev/sda. Eventually this should be usable too when we need to track back from
a mapper device as well, but for now this is left out to be simpler and quicker
to mock and use in tests.

This also in effect removes the currently entirely unused and unclear use-case
around getting a Disk() for a non-physical mapper volume device node, where the
backing device of the mapper device is not a physical disk. These cases
probably exist in real life on systems, but we do not need to concern ourselves
with them for now and it is becoming ever more clear that the disks package is
about real physical disks at the end of the day, so supporting these
non-physical disks is out of scope of these functions.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>