tests/nested/manual: add regression test with old/existing initrd and new snapd
This test ensures that new snapd when doing resealing on a system with an "old"
initrd (i.e. one that is currently published and not from the PR), we are able
to come back after the reboot to try the new kernel after resealing.

This test has two variants, one which starts with the new snapd and one which
starts with the old/existing snapd on the stable channel.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>