tests/main/security-dev-input-event-denied: fix the test
Now that device cgroup assignment when no devices match is done properly, i.e.
we do not end up in a cgroup with just the common set of devices if none are
assigned, the test needs to be updated as we correctly observe first the
AppArmor denial, and then a device cgroup one.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>