i/builtin: suppress denials of sys_module capability
The rtnetlink kernel interface can trigger a code path where the
sys_module capability is being checked, but modern applications using
the network-control interface do not really need it. So, let the
network-control interface suppress the denials on this capability,
whereas the kernel-module-control, which adds it to the allowed set,
needs to declare its usage in order to avoid the denial rule to be
emitted.