bootloader: add support for piboot
Add support for the official RPi bootloader, so we can use it instead
of RPi's U-Boot port. The U-Boot port is a community effort, so it has
a series of problems, like:

  1. It takes some time until it gets ported when a new RPi model
     appears in the market
  2. It lacks support for USB drivers so USB is not possible

This commit makes it possible to use RPi bootloader, however there are
some limitations that cannot be directly solved as this firmware is
closed source:

  1. We can cold-boot only from the first partition in the disk. This
     implies that we need to write boot assets to the ubuntu-seed
     partition instead of to ubuntu-boot.
  2. The OS updates mechanism depends on a volatile flag that gets
     removed in cold boots. That makes it not possible to distinguish
     sometimes between failed updates and having power-cycled a device
     before really trying a pending update.
  3. There is no scripting language for the RPi bootloader. The only
     way to influence its behavior is by changes to the
     {config,tryboot}.txt files.

The implementation leverages the os_prefix [1] setting in the
bootloader configuration files to select the
kernel/initramfs/dtb/dtbos/cmdline to use in the next
boot. Environment is stored in key=value pairs in text files that are
translated to bootloader configuration when needed (a new kernel is
installed, the run mode changes, etc.). To be able to try new kernels,
fail-safe OS updates are used [2].

[1] https://www.raspberrypi.com/documentation/computers/config_txt.html#os_prefix
[2] https://www.raspberrypi.com/documentation/computers/raspberry-pi.html#fail-safe-os-updates-tryboot