o/snapstate: improve install/update tests (#10850)
I've noticed that it's a bit cumbersome to debug behaviour related to retrying and tasks in the install and update tests. First, most tests lock the state at the beginning and defer an unlock but then unlock again before the call to Settle(). However, if the call to Settle() fails then the test panics because the deferred statement unlocks and already unlocked lock. This hides the true failure and forces us to change the test and re-run it to get the real failure. Second, if the test fails to Settle(), it just logs that the settle failed to converge. There's no information on which tasks were running and their statuses.
Mainly this PR changes the settle helper to do an unlock/defer lock on its own.