kernel/fde: add DeviceUnlockKernelHookDeviceMapperHandler, use w/ disks pkg
Add two new functions, IsEncryptedDeviceMapperName to identify if a device
mapper node is with a device-unlock encrypted mapper name device, and
DeviceUnlockKernelHookDeviceMapperHandler to combine the identification with
the task of "un-mapping" the device mapper to the original source disk. Also
register this with the disks package so users of the disks package will get
this built-in when they also import the fde package.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>